requirements.txt
# Hello Daft - Python Dependencies
# 核心依赖

# Daft - 分布式数据框架
getdaft>=0.2.0

# Ray - 分布式计算框架
ray[default]>=2.9.0

# LanceDB - 向量数据库
lancedb>=0.4.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0

# 机器学习和嵌入
sentence-transformers>=2.2.0
torch>=2.0.0

# Jupyter 和可视化
jupyter>=1.0.0
notebook>=7.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# 工具库
tqdm>=4.65.0
python-dotenv>=1.0.0
pyyaml>=6.0

# 数据生成和测试
faker>=18.0.0
pytest>=7.3.0
pytest-cov>=4.1.0

# Kubernetes 客户端（可选）
kubernetes>=26.1.0

# 对象存储（可选）
boto3>=1.26.0  # AWS S3
minio>=7.1.0   # MinIO

# 监控和日志（可选）
prometheus-client>=0.16.0
loguru>=0.7.0
